many updates to PeerList, and SelectModeScreen
