+++ src/components/PeerList/index.js
@@ -1,0 +1,1 @@
+export {PeerList} from './PeerList';
